backend/learning_module/utils/learner_chat.py
from google import genai
from google.genai import types
from dotenv import load_dotenv
import os

load_dotenv()
api_key = os.getenv('GEMINI_API_KEY')
print(api_key)

client = genai.Client(api_key=api_key)

