backend/learning_module/requirements.txt
youtube_transcript_api
python-dotenv
pathlib
urllib3
google
google-genai
Flask
flask-cors
Pillow
werkzeug